lf builtins: move get_builtin_prompt to prompts.py, fix flow imports
Move get_builtin_prompt from __init__.py to dedicated prompts.py module.
Add explicit Flow imports to builtin flows instead of relying on noqa.
Add voice field to Synthesize for voice-specific synthesis.
